Improve logging (#1907)
Make it easier for users to filter log messages.

**breaking changes**: 

For Python:
Don't print messages to stdout, but collect them in `ReturnData` and forward them to python logging, making it easier to filter specific messages or to disable output completely.

For C++:
Don't print messages to stdout, but collect them in `ReturnData`, and let the user decide what to do with it. 

This implies, that messages will no longer be shown as they occur, but only after `runAmiciSimulation` finishes. Downside: no live output. Upside: We can prefix messages with simulation condition IDs, making it easier to see which where any errors occurred. This is particularly helpful for `runAmiciSimulation` with multiple conditions. 

Also resolves #1550.

Note: This should also allow us to stop using `wurlitzer` and should resolve any weird output redirection issues that occurred previously. To be done separately.


Co-authored-by: Fabian Fröhlich <fabian@schaluck.com>