Fix ReturnData::{preeq_wrms,posteq_wrms} with FSA and check_sensi_steadystate_conv_=True (#2071)
Documentation says, that ReturnData::{preeq_wrms,posteq_wrms} refers to the `weighted root-mean-square of the rhs when steadystate was reached`, independent of `Solver::check_sensi_steadystate_conv_`, but this wasn't what happened. Now it is.

Closes #2066 

Furthermore, this prevents unnecessary simulation if we are already in a steadystate when entering `SteadystateProblem::runSteadystateSimulation`.