SBML import: Support for rateOf (#2120)
Adds support for SBML's `rateOf(.)` for most not-so-exotic cases.

We currently can't check for functions with `<csymbol encoding="text" definitionURL="http://www.sbml.org/sbml/symbols/rateOf">`, and only check for functions named `rateOf`. This should be safe as long as we use the `SBMLFunctionDefinitionConverter` before. 

Tested in the following SBML semantic test suite cases: 01248,01249,01250,01251,01252,01253,01254,01255,01256,01257,01258,01259,01260,01261,01262,01263,01264,01265,01266,01267,01268,01269,01270,01293,01294,01295,01296,01297,01298,01299,01321,01322,01400,01401,01402,01403,01405,01406,01408,01409,01455,01456,01457,01458,01459,01460,01461,01462,01463,01482,01483,01525,01526,01527,01528,01529,01540,01541,01542,01543

Closes #769