Remove setuptools<64 restriction (#2180)
We had to require setuptools<64 because it change its temporary build directory from a known path to a randomly named temporary directory. However, setuptools<64 doesn't work with for Python3.12 (#2179). Therefore, this adds an option to change the setuptools temporary build directory via the `AMICI_BUILD_TEMP` environment variable, so we have those temp files in a known path for coverage analysis. Not meant to be used by users.