Allow subsetting simulation conditions in simulate_petab
simulate_petab was mainly intended to simulate all conditions at once,
but the `simulation_conditions` argument provides also an easy way to
simulate only a subset of conditions. However, this currently fails,
because the full parameter mapping is used an then wrong conditions
are assigned.

This is fixed here. Also it saved a minuscule amount of time.