Don't eliminate parameters that are initial assignment targets (pt2)
Currently, parameters that are targets of initial assignments don't show up as parameters or expressions in the amici model. This is rather not what most users would expect.

Therefore, treat all SBML parameters that are initial assignment targets and whose initial assignment does not evaluate to a number (for those that do, see #2304) as amici expressions.
Those static expressions will be handled more efficiently after #2303.

Related to #2150.