Don't eliminate parameters that are initial assignment targets (pt1)
Currently, parameters that are targets of initial assignments don't show up as parameters or expressions in the amici model.
This is rather not what most users would expect.

As a first step: treat all SBML parameters that are initial assignment targets and whose initial assignment evaluates to a number as amici parameters.

Related to #2150.