Remove deprecated functionality for PEtab import from individual files instead of petab.Problem
Since almost two years, it's possible to pass a `petab.Problem`, which is safer
and more convenient.
Omitting the parameter table for model import (since there is no option to supply
this) is likely to produce unwanted results (see #2458 and #2455 for more details),
and therefore, this functionality is best removed.