PEtab import: Fix potentially incorrect sensitivities with state-dependent sigmas
Due to PEtab not allowing and observables in noiseFormula and
AMICI not allowing state variables in sigma expressions, we are trying
to replace any state variables by matching observables during PEtab import.

However, if there were multiple observables with the same observableFormula,
or one observableFormula was contained in another one, the substition
of observableFormula for observableId could - depending on the ordering -
result in noiseFormulas depending on other observables, which could lead to
incorrect sensivities due to https://github.com/AMICI-dev/AMICI/issues/2561.

This change ensures that we are only using the observableId from the same
row of the observable table as the noiseFormula.

---------

Co-authored-by: Dilan Pathirana <59329744+dilpath@users.noreply.github.com>