Improve model extension import (#2604)
Previously, it wasn't possible to import two model modules with the same name. Now this is at least possible if they are in different locations. Overwriting and importing a previously imported extension is still not supported.

In essence, we avoid using `import` and `sys.modules`, and directly import the specific files. For a single file, this works pretty neatly, but for an actual package less so. 

Closes #1936