Add `amici::Event` class
For more comprehensive SBML event handling, we need track additional attributes of event triggers and assignments.
This will be easiest if we encapsulate event properties in a dedicated class.
This PR adds a basic Event class that stores the initial value of the root function.
This will be extended shortly.