Refactor SBML event processing, defer computing state updates (#2742)
To support `use_values_from_trigger_time=true`, `deltax` may have to be
computed from `x_old` instead of `x`.
At the stage of processing the SBML model, there is no concept of `x_old`.
Thus, the state updates should only be computed inside the `DEModel`.
Also independently of `use_values_from_trigger_time`, this seems to be the
more appropriate place to compute the bolus.
Therefore, store only the event assignment in `Event`, and compute the bolus
on demand.