Fix `EventHandlingSimulator::get_simulation_state` (#2901)
Previously, `get_simulation_state` not only created a `SimulationState` but also updated the current workspace.
This is unexpected and causes problems when handling nested events when the workspace is updated but the solver was not reinitialized yet (kind of reverting the primary event assignment).
Thus, decouple snapshotting of the current state from pulling Solver state.