Refactor rateOf handling (#3036)
* doc: Skip petab-sciml for now

GitHub Action runners run out of disk space when installing petab-sciml with all its huge dependencies.
Don't install that for now. So far, it's not used anywhere for the documentation build as far as I can see.
This won't prevent enabling intersphinx later on.

* Refactor rateOf handling

Pull rateOf-handling out of DEModel and keep it along other SBML processing where it belongs.

This is easier to follow and prevents some lingering issues with the old approach due to the xdot / w interdependencies.

This also handles rateOf expressions in some additional, previously unsupported places like event assignments.